Update Dijkstra.py
Trying to figure out why it isn't printing all the lists. Found it- it cuts the number of paths in half when we change the specified. By printing the value of height_img and width_img, I learned that they counted an extra column and row. So in a 3x3, it waid there were 3 rows and 3 columns,, which included 0, so actually 4
Hence to fix, I simply subtracted 1 from the height and with values in the list comprehension